changed the DB creation script so the lat and long values are ints and not floats

+++ src/eltoraz/pug/server/DBcreation.sql
@@ -26,8 +26,8 @@ CREATE TABLE IF NOT EXISTS `users` (
 CREATE TABLE IF NOT EXISTS `locations` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `name` varchar(50) DEFAULT NULL,
-  `lat` float NOT NULL,
-  `longi` float NOT NULL,
+  `lat` int(11) NOT NULL,
+  `longi` int(11) NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;
